backend/pyproject.toml
[tool.poetry]
name = "ghostcall-backend"
version = "0.1.0"
description = "Agent task marketplace backend"
authors = ["ghostcall"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.0"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.43"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
